cod/es.uca.gii.csi2223.richard.check/src/util/Database.java
package util;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.io.IOException; 

import java.lang.reflect.InvocationTargetException;
import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.Properties;

public class Database {

	public static String getPropertiesUrl() { return "./db.properties"; }
	
	public static Connection Connection() throws SQLException, IOException {

		Properties properties = Config.Properties(getPropertiesUrl());
			
		return DriverManager.getConnection(
				properties.getProperty("jdbc.url"),
				properties.getProperty("jdbc.username"),
				properties.getProperty("jdbc.password"));
	}
	
	public static void LoadDriver() throws InstantiationException, IllegalAccessException, 
										   ClassNotFoundException, IOException, 
										   IllegalArgumentException, InvocationTargetException, 
										   NoSuchMethodException, SecurityException {
		
		Class.forName(Config.Properties(getPropertiesUrl()).getProperty(
				"jdbc.driverClassName")).getDeclaredConstructor().newInstance();
	}
	
	public static String String2Sql(String s, boolean bAddQuotes, boolean bAddWildcards) {
		
		s = s.replace("'", "''");
		if(bAddWildcards) s = "%" + s + "%";
		if(bAddQuotes) s = "'" + s + "'";
		
		return s;
	} 
	
	public static int Boolean2Sql(Boolean b) {
		
		if(b) return 1;
		else return 0;		
	}
	
	public static int LastId (Connection con) throws IOException, SQLException{
		
		ResultSet rs = null; 
		Properties properties = Config.Properties(getPropertiesUrl());
		
		try {
			rs = con.createStatement().executeQuery(properties.getProperty("jdbc.lastIdSentence"));			
			 
			rs.next();
			return rs.getInt(1);
		} 
		catch (SQLException e) { throw e; }
		finally { if (rs != null) rs.close(); }
	}
}